+++ src/actions/__tests__/test-MessageFormActions.js
@@ -9,6 +9,7 @@ describe('MessageFormActions', () => {
             ['update', 'UPDATE', {}],
             ['remove', 'REMOVE', 'id'],
             ['whisperTo', 'WHISPER_TO', 'to'],
+            ['appendFile', 'APPEND_FILE', 'id', 'name.png', {}],
         ].forEach(([akey, ckey, ...args]) => {
             expect(Actions[akey](...args).type)
                 .toEqual(ACTIONS[ckey]);