+++ src/server/models/__tests__/test-icon.js
@@ -1,0 +1,10 @@
+describe('Icon', () => {
+    const {Model} = require('../model');
+
+    jest.dontMock('../icon');
+    require('../icon');
+
+    it('inherits Model', () => {
+        expect(Model).toBeCalledWith('icons');
+    });
+});